added

+++ CodeForces/O_Fibonacci.cpp
@@ -5,7 +5,7 @@ int main()
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
     // --- your code here ---
-    int n;
+    long long n;
     cin >> n;
     if (n == 1)
     {
@@ -18,10 +18,10 @@ int main()
         return 0;
     }
 
-    int prev2 = 0;
-    int prev1 = 1;
-    int cur = 0;
-    for (int i = 3; i <= n; i++)
+    long long prev2 = 0;
+    long long prev1 = 1;
+    long long cur = 0;
+    for (long long i = 3; i <= n; i++)
     {
         cur = prev2 + prev1;
         prev2 = prev1;
